Add variable intra-annual CostVariable feature
This commit adds a feature to vary technology variable costs at the
time step level. Prior to this commit, technology variable costs were
indexed over period & vintage pairs only -- i.e. it was not possible to vary
the variable cost throughout the year.

This is accomplished by adding an additional table to the SQL/SQLite
files: CostVariableVariable. The CostVariableVariable parameter is indexed
over {region, technology, season, and time of day} whereas the CostVariable
parameter is indexed over {region, period, technology, and vintage}.

Notice the CostVariableVariable parameter is not indexed over period or vintage.
This means that in every period (i.e. year), all vintages of the technology are
assumed to have the same intra-annual variation in variable cost.

The default value for the CostVariableVariable parameter is 1.

The implementation works as follows: If you would like to have technology
A's variable cost vary throughout the year, begin by inputting its nominal
(or annual average) variable cost in the CostVariable table as before. Then,
use the CostVariableVariable table to input factors for each time step of
the year that, when multiplied by the technology's nominal variable cost,
yield the desired variable cost..

For instance, say the variable cost in question is the cost to import electricity
from an export market. You might put the nominal variable cost at 0.15 $/kWh, which
represents the annual average. Then, using the CostVariableVariable parameter for the
same technology, you might place values of 1.5 for time steps during the winter and
values of 0.8 for time steps during the summer. This would result in the price of
imported electricity being increased by 50% in winter hours, decreased by 20% in
summer hours, and kept the same during spring and fall hours.

Important: this feature does not apply to technologies in the "tech_annual"
set -- i.e. technologies with constant annual output (see the Temoa documentation
for more information). These technologies assume their nominal variable cost in all
time steps.